Remove duplicated files
